requirements.txt
panda
yfinance
yahoofinancials